third_party/blink/web_tests/platform/webview/external/wpt/pointerevents/pointerlock/pointerevent_pointerlock_supercedes_capture-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Unhandled rejection: If you see this error we have a bug. Please report this bug to chromium.
[FAIL] mouse no pointercapture while pointerlock
  assert_equals: Pointer is unlocked expected 2 but got 0
Harness: the test ran to completion.
